requirements.txt
fastapi
uvicorn
httpx
hazelcast-python-client
psycopg2-binary